+++ crates/cfxcore/executor/src/state/state_object/mod.rs
@@ -66,9 +66,8 @@ use super::{
     overlay_account::{AccountEntry, OverlayAccount, RequireFields},
 };
 use crate::substate::Substate;
-use cfx_rpc_eth_types::StateOverride;
 use cfx_statedb::{Result as DbResult, StateDbExt, StateDbGeneric as StateDb};
-use cfx_types::{AddressWithSpace, Space};
+use cfx_types::AddressWithSpace;
 use parking_lot::RwLock;
 use std::collections::{BTreeSet, HashMap};
 
@@ -110,14 +109,6 @@ impl State {
             checkpoints: Default::default(),
             global_stat: world_stat,
         })
-    }
-
-    pub fn new_with_override(
-        db: StateDb, state_override: &StateOverride, space: Space,
-    ) -> DbResult<Self> {
-        let mut state = Self::new(db)?;
-        state.apply_override(state_override, space)?;
-        Ok(state)
     }
 
     pub fn prefetch_accounts(
