+++ crates/cfxcore/executor/src/state/state_object/state_override.rs
@@ -1,13 +1,21 @@
 use super::State;
 use crate::state::overlay_account::AccountEntry;
 use cfx_rpc_eth_types::StateOverride;
-use cfx_statedb::{Result as DbResult, StateDbExt};
+use cfx_statedb::{Result as DbResult, StateDbExt, StateDbGeneric as StateDb};
 use cfx_types::{AddressWithSpace, Space};
 
 /// Apply the state override to the state object, only used for rpc call eg
 /// eth_call, eth_estimateGas etc.
 impl State {
-    pub(super) fn apply_override(
+    pub fn new_with_override(
+        db: StateDb, state_override: &StateOverride, space: Space,
+    ) -> DbResult<Self> {
+        let mut state = Self::new(db)?;
+        state.apply_override(state_override, space)?;
+        Ok(state)
+    }
+
+    fn apply_override(
         &mut self, state_override: &StateOverride, space: Space,
     ) -> DbResult<()> {
         assert!(self.checkpoints.read().is_empty());
